requirements.txt
alphashape==1.3.1
anndata==0.9.2
descartes==1.1.0
gseapy==0.10.8
h5py==3.6.0
imageio==2.9.0
jupyter==1.0.0
matplotlib>=3.1.0,<3.7
networkx==2.7.1
natsort==8.1.0
notebook==6.4.8
numba==0.55.1
numpy==1.21.2
pandas>=0.25.0,<=1.4.1
pynndescent==0.5.6
pysocks==1.7.1
PyYAML==6.0.1
plotly==5.6.0
scikit-learn>=0.20.0,<=1.0.2
scanpy==1.8.2
scipy==1.7.3
seaborn==0.11.2
setuptools==58.0.4
shapely==1.7.1
sinfo==0.3.4
stdlib-list==0.8.0
statsmodels==0.14.0
tables==3.6.1
torch>=1.1.0
tqdm==4.63.1
umap-learn>=0.4.1,<=0.5.2
wheel==0.21.0